+++ tests/testthat/test-SHARED-getAdjacency-propd.R
@@ -7,18 +7,18 @@ data(crabs)
 x <- crabs[,4:8]  # data matrix with 5 variables
 y <- crabs[,1]    # group vector
 
-test_that("getAdjFDR works properly for theta", {
+test_that("getAdjacencyFDR works properly for theta", {
 
     # get propd object
     pr <- propd(x, as.character(y), p=10)
     pr <- updateCutoffs(pr, number_of_cutoffs=10)
 
     # get adjacency matrix
-    adj <- getAdjFDR(pr)
+    adj <- getAdjacencyFDR(pr)
 
     # get expected adjacency matrix
     adj_expected <- matrix(0, nrow = ncol(x), ncol = ncol(x))
-    adj_expected[propr:::get_theta_matrix(pr) <= getCutoffFDR(pr)] <- 1
+    adj_expected[propr:::getMatrix(pr) <= getCutoffFDR(pr)] <- 1
     adj_expected[diag(adj_expected)] <- 1
     rownames(adj_expected) <- colnames(x)
     colnames(adj_expected) <- colnames(x)
@@ -27,14 +27,14 @@ test_that("getAdjFDR works properly for theta", {
     expect_equal(adj, adj_expected)
 })
 
-test_that("getAdjFDR and getSignificantResultsFDR return coherent results", {
+test_that("getAdjacencyFDR and getSignificantResultsFDR return coherent results", {
 
     # get propd object
     pr <- propd(x, as.character(y), p=10)
     pr <- updateCutoffs(pr, number_of_cutoffs=10)
 
     # get adjacency matrix
-    adj <- getAdjFDR(pr)
+    adj <- getAdjacencyFDR(pr)
 
     # get significant results
     results <- getSignificantResultsFDR(pr)
@@ -45,7 +45,7 @@ test_that("getAdjFDR and getSignificantResultsFDR return coherent results", {
     }
 })
 
-test_that("getAdjFstat works properly", {
+test_that("getAdjacencyFstat works properly", {
 
     for (fdr in c(TRUE, FALSE)){
 
@@ -54,11 +54,11 @@ test_that("getAdjFstat works properly", {
         pr <- updateF(pr)
 
         # get adjacency matrix
-        adj <- getAdjFstat(pr, fdr=fdr)
+        adj <- getAdjacencyFstat(pr, fdr=fdr)
 
         # get expected adjacency matrix
         adj_expected <- matrix(0, nrow = ncol(x), ncol = ncol(x))
-        adj_expected[propr:::get_theta_matrix(pr) <= getCutoffFstat(pr, fdr=fdr)] <- 1
+        adj_expected[propr:::getMatrix(pr) <= getCutoffFstat(pr, fdr=fdr)] <- 1
         adj_expected[diag(adj_expected)] <- 1
         rownames(adj_expected) <- colnames(x)
         colnames(adj_expected) <- colnames(x)
@@ -68,7 +68,7 @@ test_that("getAdjFstat works properly", {
     }
 })
 
-test_that("getAdjFstat and getSignificantResultsFstat return coherent results", {
+test_that("getAdjacencyFstat and getSignificantResultsFstat return coherent results", {
 
     for (fdr in c(TRUE, FALSE)){
 
@@ -77,7 +77,7 @@ test_that("getAdjFstat and getSignificantResultsFstat return coherent results", 
         pr <- updateF(pr)
 
         # get adjacency matrix
-        adj <- getAdjFstat(pr, fdr=fdr)
+        adj <- getAdjacencyFstat(pr, fdr=fdr)
 
         # get significant results
         results <- getSignificantResultsFstat(pr, fdr=fdr)